Move Selenium based tests into `selenium_tests.py` files and run them in a separate travis build
